Inserido o 'alert' para confirmar o delete

+++ cad_usu_dados.php
@@ -502,7 +502,7 @@ endwhile;
                                             </div>
                                         </div>
                                     </div>
-					<a href="usu_delete.php?id=<?php echo $idusuario; ?>"
+					<a href="javascript: if(confirm('Tem certeza que deseja deletar o usuário <?php echo($nomeguerra); ?>?')) location.href='usu_delete.php?id=<?php echo $idusuario; ?>';"
 						<button class="btn btn-danger">
                                                 	Deletar usuário
                                        		</button>
